(svn r134) -Fix: [976583] parent_list was too small -Fix: [981934] Memoryleak in parent_list -Add: General protection around Sprites

+++ viewport.h
@@ -24,7 +24,7 @@ void UpdateViewportPosition(Window *w);
 void OffsetGroundSprite(int x, int y);
 
 void DrawGroundSprite(uint32 image);
-void DrawGroundSpriteAt(uint32 image, int x, int y, byte z);
+void DrawGroundSpriteAt(uint32 image, int16 x, int16 y, byte z);
 void AddSortableSpriteToDraw(uint32 image, int x, int y, int w, int h, byte dz, byte z);
 void *AddStringToDraw(int x, int y, StringID string, uint32 params_1, uint32 params_2);
 void AddChildSpriteScreen(uint32 image, int x, int y);
